Small bugfix regarding the thumbnail.

+++ src/main/java/org/knipsX/model/picturemanagement/Picture.java
@@ -114,10 +114,10 @@ public class Picture extends Observable implements PictureInterface {
      */
     public String getThumbnailPath() {
         String id = ""+this.hashCode();
-        if(!this.isTempFileCreated) {
+        if(!this.isTempFileCreated && this.bigThumbnail != null) {
             try {
                 File tempFile = File.createTempFile(id, ".jpg");
-                ImageIO.write((RenderedImage) this.getBigThumbnail(), "jpg", tempFile);
+                ImageIO.write((RenderedImage) this.bigThumbnail, "jpg", tempFile);
                 tempFile.deleteOnExit();
                 this.tempFilePath = tempFile.getAbsolutePath();
                 this.isTempFileCreated = true;
@@ -245,7 +245,6 @@ public class Picture extends Observable implements PictureInterface {
      * @see org.knipsX.model.picturemanagement.PictureInterface#getBigThumbnail()
      */
     public Image getBigThumbnail() {
-        this.initThumbnails();
         return this.bigThumbnail;
     }
